[RENAME] SecureBiometricDevice to SecureBiometricInterface
