+++ secure-biometric-device/example/html/index.html
@@ -3,7 +3,7 @@
 
 <head>
     <meta charset="UTF-8">
-    <title>Secure Biometric Device</title>
+    <title>Secure Biometric Interface</title>
 
     <script src="../../dist/iife/index.js"></script>
 </head>
