+++ storybook-example/stories/SecureBiometricDevice.stories.js
@@ -33,7 +33,7 @@ const sbiEnv = {
 };
 
 export default {
-  title: "Secure Biometric Device",
+  title: "Secure Biometric Interface",
   tags: ["autodocs"],
   render: (args) => renderComponent(args),
   argTypes: {
@@ -132,7 +132,7 @@ const renderComponent = (args) => {
   );
 };
 
-export const SecureBiometricDeviceStory = {
+export const SecureBiometricInterfaceStory = {
   args: {
     buttonLabel: "scan_and_verify",
     disable: false,
@@ -144,14 +144,14 @@ export const SecureBiometricDeviceStory = {
   },
 };
 
-SecureBiometricDeviceStory.parameters = {
+SecureBiometricInterfaceStory.parameters = {
   docs: {
     source: {
       code: `
       <div id="secure-biometric-device"></div>
 
       // in javascript
-      SecureBiometricDevice.init({
+      SecureBiometricInterface.init({
         container: document.getElementById("secure-biometric-device"),
         buttonLabel: "scan_and_verify",
         sbiEnv: {
