+++ storybook-example/stories/index.ts
@@ -4,7 +4,7 @@ import {
   ISignInWithEsignetProps,
   OidcConfigProp,
 } from "../../sign-in-with-esignet/src/lib/SignInWithEsignet/ISignInWithEsignetProps";
-import { init } from "./../../secure-biometric-device/lib/secureBiometricDevice";
+import { init } from "./../../secure-biometric-device/lib/secureBiometricInterface";
 import { div } from "../../secure-biometric-device/utility";
 
 export {
